Refactor:(profile_overlay 분류)
